solution 1.0 --with hook in Comments
